EditGassmonitor

+++ src/main/java/springmvc/respository/GasMonitorMapper.java
@@ -21,7 +21,7 @@ public class GasMonitorMapper implements RowMapper<GasMonitor> {
     public GasMonitor mapRow(ResultSet rs, int i) throws SQLException {
         GasMonitor g = new GasMonitor();
         g.setId(rs.getInt("serialnumber"));
-        g.setMaxWeight(rs.getInt("max_weight"));
+        g.setMaxWeight(rs.getDouble("max_weight"));
         g.setCustomerId(rs.getInt("cust_id"));
         g.setSupplierId(rs.getInt("supp_id"));
         g.setGasTanks(rs.getInt("number_gastanks"));
